requirements.txt
# requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
httpx==0.27.0

# AI/LLM
groq==0.11.0

# Caching
redis==5.0.8
hiredis==2.3.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0

# Web Scraping
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0

# Validation & Serialization
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0

# Environment & Utilities
python-dotenv==1.0.1

# Background Tasks (Optional)
celery==5.4.0

# Testing
pytest==8.3.0
pytest-asyncio==0.23.8

# Date utilities
python-dateutil==2.9.0

# Async Redis support
redis[hiredis]==5.0.8
